Further namespace support
User has two dashboards
 - Brands::UserDashboard
 - UserDashboard

User has one address. Not that it is not belongs to, it is has_one

Address has two dashboards
 - Brands::AddressDashboard
 - AddressDashboard

We want to show different fields of Address in the admin and in the brands

When we render
	Brands::UserDashboard
we expect that we will render
	AddressDashboard

but instead the Field::HasOne and ResourceResolved find the AddressDashboard
instead of the namespaced Brands::AddressDashboard.

This commit introduces and option to Field::HasOne.with_option(namespace: "brands")

It should be able to find it on its own, but I have to think about how to do it.